+++ prescia/plugins/bi_groups/module.php
@@ -34,8 +34,8 @@ class mod_bi_groups extends CscriptedModule  {
 					if ($action == CONS_ACTION_INCLUDE) $data['level'] = 99;
 					else unset($data['level']);
 				}
-			} else if ($action == CONS_ACTION_UPDATE) {
-				$lvl = $this->parent->dbo->fetch("SELECT level FROM ".$baseModule->dbname." WHERE id=".$data['id']);
+				} else if ($action == CONS_ACTION_UPDATE) {
+					$lvl = $this->parent->dbo->fetchPrepared("SELECT level FROM ".$baseModule->dbname." WHERE id=?", 'i', array((int)$data['id']));
 				if ($lvl == 100 && $_SESSION[CONS_SESSION_ACCESS_LEVEL] < 100)  {
 					$this->parent->log[] = $this->parent->langOut("cannot_change_master_group");
 					$this->parent->setLog(CONS_LOGGING_WARNING);
@@ -52,9 +52,9 @@ class mod_bi_groups extends CscriptedModule  {
 				$this->parent->setLog(CONS_LOGGING_WARNING);
 				$data['level'] = 99;
 			}
-		} else if ($action == CONS_ACTION_DELETE && $this->parent->safety) {
-			# gets the group level, if it's higher, disallow delete
-			$lvl = $this->parent->dbo->fetch("SELECT level FROM ".$baseModule->dbname." WHERE id=".$data['id']);
+			} else if ($action == CONS_ACTION_DELETE && $this->parent->safety) {
+				# gets the group level, if it's higher, disallow delete
+				$lvl = $this->parent->dbo->fetchPrepared("SELECT level FROM ".$baseModule->dbname." WHERE id=?", 'i', array((int)$data['id']));
 			if ($lvl >= $_SESSION[CONS_SESSION_ACCESS_LEVEL]) {
 				$this->parent->log[] = $this->parent->langOut("group_cannot_change_higher");
 				$this->parent->setLog(CONS_LOGGING_WARNING);